+++ packages/studio-components/src/Utils/index.tsx
@@ -231,7 +231,7 @@ export {
 } from './parseCSV';
 export { inferredGraphFields } from './inferredGraphFields';
 export { parseJSON } from './parseJSON';
-export { parseSQL } from './parseSQL';
+
 export { extractProperties } from './inferredSchema';
 export { handleExpand, uniqueElementsBy } from './expand';
 export { transSchema } from './transSchema';