+++ packages/studio-graph/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphscope/studio-graph
 
+## 0.1.16
+
+### Patch Changes
+
+- update website
+- Updated dependencies
+  - @graphscope/studio-components@0.1.16
+  - @graphscope/use-zustand@0.1.8
+
 ## 0.1.15
 
 ### Patch Changes